Consolidate List and Recipient forms, removed RecipientForm
