+++ client/src/components/index.js
@@ -25,7 +25,6 @@ import ExistingClassCard from './classesPage/components/ExistingClassCard';
 import NewClassCard from './classesPage/components/NewClassCard';
 import CampaignForm from './classesPage/forms/CampaignForm';
 import ListForm from './classesPage/forms/ListForm';
-import RecipientForm from './classesPage/forms/RecipientForm';
 import SelectionForm from './classesPage/forms/SelectionForm';
 
 /* REFRESHRS */
@@ -61,7 +60,6 @@ export {
   NewClassCard,
   CampaignForm,
   ListForm,
-  RecipientForm,
   SelectionForm,
   RefreshrListView,
   RefreshrCard,
