+++ client/src/theme.js
@@ -5,28 +5,29 @@ import { createMuiTheme } from '@material-ui/core/styles';
 //#4C99E5 Ocean blue(Font)
 //#FFFFFF white(card color)(nav)
 
-  //breakpoint values {xs: 0, sm: 600, md: 960, lg: 1280, xl: 1920}
+//breakpoint values {xs: 0, sm: 600, md: 960, lg: 1280, xl: 1920}
 
 export default createMuiTheme({
-    typography: {
-        htmlFontSize: 10,
-        useNextVariants: true,
-        body1: {color: "inherit"}
+  typography: {
+    htmlFontSize: 10,
+    useNextVariants: true,
+    body1: { color: "inherit" },
+    body2: { color: "inherit" }
+  },
+  palette: {
+    primary: {
+      main: '#0b2742',
+      contrastText: '#FFFFFF'
     },
-    palette: {
-        primary: {
-            main: '#0b2742',
-            contrastText: '#FFFFFF'
-        },
-        secondary: {
-            main: '#FFFFFF',
-            contrastText: '#0b2742'
+    secondary: {
+      main: '#FFFFFF',
+      contrastText: '#0b2742'
 
-        },
-        background: {
-         default: '#0b2742', 
-         paper: '#0b2742'  
-        },
-        tonalOffset: 0.2
-    },  
+    },
+    background: {
+      default: '#0b2742',
+      paper: '#0b2742'
+    },
+    tonalOffset: 0.2
+  },
 })
